A few bugfixes and improvements
First, the TEXT_FINDER_XPATH was relaxed to permit the algorithm to
capture text in small HTML snippets, specially malformed HTML.

The "get_xpath_frequencydistribution" was having arbitrary behaviour.
When two or more parent paths had the same number of appearances in the
HTML tree, one of them was beign selected at random. That's because the
function relied on "collections.Counter.most_common" to sort the paths
in a list, and, as documented, "elements with equal counts are ordered
arbitrarily".

  https://docs.python.org/3/library/collections.html#collections.Counter.most_common

In practice, the algorithm was producing different results when ran
multiple times over the same input data.

In "calc_across_paths_textnodes", nodes containing only blank lines are
not accounted for. Also, leading and trailing whitespace does not
compute in the total string length of a node.

The high pass filter was not working when the HTML presented a single
text node. The list comprehension that selects text nodes for extraction
required that the node had a total string length (TSL) larger than the
global average (AABSL).

    filtered = [parpath_tnodes for parpath_tnodes in pars_tnodes
                if parpath_tnodes[1][2] > avg]

When the tree contained a single text node, that node's total length was
always equal to the global average, and the node would never be
selected. The test was replaced with a "greater than or equal" to deal
with this case.